Time: 80 ms (11.17%), Space: 37.1 MB (9.9%) - LeetHub

+++ 1497-design-a-stack-with-increment-operation/1497-design-a-stack-with-increment-operation.cpp
@@ -7,7 +7,7 @@ public:
     }
     
     void push(int x) {
-        if((int)st.size()<maxSize)
+        if(st.size()<maxSize)
         {
             st.push(x);
         }
